lwio: CID# 11708, NO_EFFECT.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@52735 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lwio/utils/fileutils.c
@@ -572,7 +572,7 @@ SMBCopyFileWithPerms(
     CHAR szBuf[1024+1];
     int  iFd = -1;
     int  oFd = -1;
-    DWORD dwBytesRead = 0;
+    LONG dwBytesRead = 0;
 
     if (IsNullOrEmptyString(pszSrcPath) ||
         IsNullOrEmptyString(pszDstPath)) {
